modify check authorization from util file

+++ apis/content.py
@@ -1,3 +1,4 @@
+import logging
 import os
 from fastapi import APIRouter, status, Depends, UploadFile, HTTPException, Request
 
@@ -5,12 +6,14 @@ from sqlalchemy.orm import Session
 from sqlalchemy.sql import select, and_
 
 from datetime import datetime
+
+from apis.util import get_token_by_header, get_user_by_token, check_authorized_by_user
 from database import get_db
 
 from boto3 import client
 
 from exceptions import GetExceptionWithStatuscode, ExceptionType
-from models.model import TrainingProgramContent, OrganizationContent, User
+from models.model import TrainingProgramContent, OrganizationContent
 from schema.content import ContentCreateResponseSchema
 
 router = APIRouter(prefix='/contents')
@@ -51,63 +54,17 @@ def upload_file_to_s3(file: UploadFile):
     return False
 
 
-def check_exist_token(request: Request):
-    headers = request.headers
-    if 'Authorization' not in headers:
-        raise GetExceptionWithStatuscode(status_code=status.HTTP_404_NOT_FOUND,
-                                         message='invalid token',
-                                         exception_type=ExceptionType.INVALID_TOKEN
-                                         )
-    return headers.get('Authorization')
-
-
-def get_user_by_token(token: str, db: Session = Depends(get_db)):
-    if not token:
-        raise GetExceptionWithStatuscode(status_code=status.HTTP_404_NOT_FOUND,
-                                         exception_type=ExceptionType.INVALID_TOKEN,
-                                         message='invalid token')
-
-    select_query = select(User).where(User.token == token)
-    user = db.scalar(select_query)
-    if not user:
-        raise GetExceptionWithStatuscode(status_code=status.HTTP_404_NOT_FOUND,
-                                         exception_type=ExceptionType.NOT_MATCHED,
-                                         message='there is no user')
-    return user
-
-def get_authorized_user_by_token(token: str, db: Session = Depends(get_db)):
-    if not token:
-        raise GetExceptionWithStatuscode(status_code=status.HTTP_404_NOT_FOUND,
-                                         exception_type=ExceptionType.INVALID_TOKEN,
-                                         message='invalid token')
-
-    select_query = select(User).where(User.token == token)
-    user = db.scalar(select_query)
-    if not user:
-        raise GetExceptionWithStatuscode(status_code=status.HTTP_404_NOT_FOUND,
-                                         exception_type=ExceptionType.NOT_MATCHED,
-                                         message='there is no user')
-    elif user.user_role_id != 3:
-        raise GetExceptionWithStatuscode(status_code=status.HTTP_401_UNAUTHORIZED,
-                                         exception_type=ExceptionType.INVALID_PERMISSION,
-                                         message='no authorization')
-
-    return user
-
-
-@router.post('/training/{training_program_id}', status_code=status.HTTP_201_CREATED, response_model=ContentCreateResponseSchema)
-async def create_training_content(request: Request, content: UploadFile, training_program_id: int, db: Session = Depends(get_db)):
+@router.post('/training/{training_program_id}', status_code=status.HTTP_201_CREATED,
+             response_model=ContentCreateResponseSchema)
+async def create_training_content(request: Request, content: UploadFile, training_program_id: int,
+                                  db: Session = Depends(get_db)):
     try:
-        token = check_exist_token(request)
-        get_authorized_user_by_token(token, db)
+        token = get_token_by_header(request)
+        user = get_user_by_token(token, db)
+        check_authorized_by_user(user)
     except GetExceptionWithStatuscode as e:
-        if e.exception_type == ExceptionType.INVALID_PERMISSION:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.NOT_MATCHED:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.INVALID_TOKEN:
-            raise HTTPException(e.status_code, e.message)
-        return
+        logging.error(e)
+        raise HTTPException(status_code=e.status_code, detail=e.message)
 
     # upload file to s3
     s3_key = upload_file_to_s3(content)
@@ -115,7 +72,8 @@ async def create_training_content(request: Request, content: UploadFile, trainin
         raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail='could not upload file')
 
     # db insert
-    training_content = TrainingProgramContent(s3_key=s3_key, file_name=content.filename, training_program_id=training_program_id)
+    training_content = TrainingProgramContent(s3_key=s3_key, file_name=content.filename,
+                                              training_program_id=training_program_id)
     db.add(training_content)
     db.commit()
     db.refresh(training_content)
@@ -136,28 +94,22 @@ def check_exist_training_content(content_id: int, db: Session = Depends(get_db))
 @router.get('/training/{content_id}', response_model=ContentCreateResponseSchema)
 async def get_training_content(request: Request, content_id: int, db: Session = Depends(get_db)):
     try:
-        token = check_exist_token(request)
+        token = get_token_by_header(request)
         get_user_by_token(token, db)
 
         training_content = check_exist_training_content(content_id, db)
         return training_content.convert_to_schema
     except GetExceptionWithStatuscode as e:
-        if e.exception_type == ExceptionType.INVALID_PERMISSION:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.NOT_MATCHED:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.INVALID_TOKEN:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.NOT_FOUND:
-            return None
-        return
+        logging.error(e)
+        raise HTTPException(status_code=e.status_code, detail=e.message)
 
 
 @router.delete('/training/{content_id}', status_code=status.HTTP_204_NO_CONTENT)
 async def delete_training_content(request: Request, content_id: int, db: Session = Depends(get_db)):
     try:
-        token = check_exist_token(request)
-        get_authorized_user_by_token(token, db)
+        token = get_token_by_header(request)
+        user = get_user_by_token(token, db)
+        check_authorized_by_user(user)
 
         training_content = check_exist_training_content(content_id, db)
         db.delete(training_content)
@@ -168,16 +120,8 @@ async def delete_training_content(request: Request, content_id: int, db: Session
         db.commit()
         return
     except GetExceptionWithStatuscode as e:
-        if e.exception_type == ExceptionType.INVALID_PERMISSION:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.NOT_MATCHED:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.INVALID_TOKEN:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.NOT_FOUND:
-            return None
-        else:
-            raise HTTPException(e.status_code, detail=e.message)
+        logging.error(e)
+        raise HTTPException(status_code=e.status_code, detail=e.message)
 
 
 @router.post('/manikin_connected/{manikin_type}', status_code=status.HTTP_201_CREATED,
@@ -185,17 +129,13 @@ async def delete_training_content(request: Request, content_id: int, db: Session
 async def create_manikin_connected(request: Request, manikin_type: str, content: UploadFile,
                                    db: Session = Depends(get_db)):
     try:
-        token = check_exist_token(request)
-        user = get_authorized_user_by_token(token, db)
+        token = get_token_by_header(request)
+        user = get_user_by_token(token, db)
+        check_authorized_by_user(user)
 
     except GetExceptionWithStatuscode as e:
-        if e.exception_type == ExceptionType.INVALID_PERMISSION:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.NOT_MATCHED:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.INVALID_TOKEN:
-            raise HTTPException(e.status_code, e.message)
-        return
+        logging.error(e)
+        raise HTTPException(status_code=e.status_code, detail=e.message)
 
     # upload file to s3
     s3_key = upload_file_to_s3(content)
@@ -216,17 +156,12 @@ async def create_manikin_connected(request: Request, manikin_type: str, content:
 @router.get('/manikin_connected/{content_id}')
 async def get_manikin_connected(request: Request, content_id: int, db: Session = Depends(get_db)):
     try:
-        token = check_exist_token(request)
+        token = get_token_by_header(request)
         user = get_user_by_token(token, db)
 
     except GetExceptionWithStatuscode as e:
-        if e.exception_type == ExceptionType.INVALID_PERMISSION:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.NOT_MATCHED:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.INVALID_TOKEN:
-            raise HTTPException(e.status_code, e.message)
-        return
+        logging.error(e)
+        raise HTTPException(status_code=e.status_code, detail=e.message)
 
     query = select(OrganizationContent).where(
         and_(OrganizationContent.id == content_id, OrganizationContent.organization_id == user.organization_id))
@@ -251,8 +186,9 @@ def check_exist_organization_content(content_id: int, db: Session = Depends(get_
 @router.delete('/manikin_connected/{content_id}', status_code=status.HTTP_204_NO_CONTENT)
 async def delete_manikin_connected(request: Request, content_id: int, db: Session = Depends(get_db)):
     try:
-        token = check_exist_token(request)
-        get_authorized_user_by_token(token, db)
+        token = get_token_by_header(request)
+        user = get_user_by_token(token, db)
+        check_authorized_by_user(user)
         organization_content = check_exist_organization_content(content_id, db)
         db.delete(organization_content)
 
@@ -262,31 +198,20 @@ async def delete_manikin_connected(request: Request, content_id: int, db: Sessio
         db.commit()
         return
     except GetExceptionWithStatuscode as e:
-        if e.exception_type == ExceptionType.INVALID_PERMISSION:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.NOT_MATCHED:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.INVALID_TOKEN:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.NOT_FOUND:
-            raise HTTPException(e.status_code, detail=e.message)
-        return
+        logging.error(e)
+        raise HTTPException(status_code=e.status_code, detail=e.message)
 
 
 @router.post('/login', status_code=status.HTTP_201_CREATED, response_model=ContentCreateResponseSchema)
 async def create_login_content(request: Request, content: UploadFile, db: Session = Depends(get_db)):
     try:
-        token = check_exist_token(request)
-        user = get_authorized_user_by_token(token, db)
+        token = get_token_by_header(request)
+        user = get_user_by_token(token, db)
+        check_authorized_by_user(user)
 
     except GetExceptionWithStatuscode as e:
-        if e.exception_type == ExceptionType.INVALID_PERMISSION:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.NOT_MATCHED:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.INVALID_TOKEN:
-            raise HTTPException(e.status_code, e.message)
-        return
+        logging.error(e)
+        raise HTTPException(status_code=e.status_code, detail=e.message)
     # upload file to s3
     s3_key = upload_file_to_s3(content)
     if not s3_key:
@@ -306,28 +231,22 @@ async def create_login_content(request: Request, content: UploadFile, db: Sessio
 @router.get('/login/{content_id}')
 async def get_login_content(request: Request, content_id: int, db: Session = Depends(get_db)):
     try:
-        token = check_exist_token(request)
+        token = get_token_by_header(request)
         get_user_by_token(token, db)
 
         login_content = check_exist_organization_content(content_id, db)
         return login_content.convert_to_schema
     except GetExceptionWithStatuscode as e:
-        if e.exception_type == ExceptionType.INVALID_PERMISSION:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.NOT_MATCHED:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.INVALID_TOKEN:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.NOT_FOUND:
-            return None
-        return
+        logging.error(e)
+        raise HTTPException(status_code=e.status_code, detail=e.message)
 
 
 @router.delete('/login/{content_id}', status_code=status.HTTP_204_NO_CONTENT)
 async def delete_login_content(request: Request, content_id: int, db: Session = Depends(get_db)):
     try:
-        token = check_exist_token(request)
-        get_authorized_user_by_token(token, db)
+        token = get_token_by_header(request)
+        user = get_user_by_token(token, db)
+        check_authorized_by_user(user)
 
         organization_content = check_exist_organization_content(content_id, db)
         db.delete(organization_content)
@@ -338,13 +257,5 @@ async def delete_login_content(request: Request, content_id: int, db: Session = 
         db.commit()
         return
     except GetExceptionWithStatuscode as e:
-        if e.exception_type == ExceptionType.INVALID_PERMISSION:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.NOT_MATCHED:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.INVALID_TOKEN:
-            raise HTTPException(e.status_code, e.message)
-        elif e.exception_type == ExceptionType.NOT_FOUND:
-            raise HTTPException(e.status_code, detail=e.message)
-
-        return
+        logging.error(e)
+        raise HTTPException(status_code=e.status_code, detail=e.message)
